NPVPN-1657: вынести хелперы подписки из роутера в app/subscription
